backend/requirements.txt
# Core dependencies
flask[async]==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1

# Database
pymongo==4.10.1

# AI/ML
google-generativeai==0.3.0
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
sentencepiece==0.1.99

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Utilities
tqdm>=4.65.0
requests>=2.31.0
python-jose==3.3.0

# Testing
pytest==8.3.4

# Optional: Development dependencies
black==23.12.1  # Code formatting
isort==5.13.2   # Import sorting
flake8==7.0.0   # Linting
